section 13 - sessions, user controller & views, register, logout, middleware for authentication

+++ routes.php
@@ -4,12 +4,19 @@ use App\Controllers\ListingController;
 
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listings/edit/{id}', 'ListingController@edit');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show'); // keep this one last so it doesn't try to match edit or create with id
 
-$router->post('/listings', 'ListingController@store');
+$router->post('/listings', 'ListingController@store', ['auth']);
 
-$router->put('/listings/{id}', 'ListingController@update');
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
 
-$router->delete('/listings/{id}', 'ListingController@destroy');
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
+
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest']);
+
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
